% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_clustering.R
\name{adjustedRandIndex}
\alias{adjustedRandIndex}
\title{culculate the adjusted Rand Index for the return class with the gd class}
\usage{
adjustedRandIndex(x, y)
}
\arguments{
\item{x}{the return class from clustering method}

\item{y}{the standard label}
}
\description{
culculate the adjusted Rand Index for the return class with the gd class
}
