% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_consboost.R
\name{cons_sampling}
\alias{cons_sampling}
\title{sampling data from full dataset to save computation resource}
\usage{
cons_sampling(object, K_cell = K_cell, K = K)
}
\arguments{
\item{object}{the FCS exprs data set after transformed}

\item{K_cell}{The number of sampling data}

\item{K}{Number of clusters}
}
\description{
sampling data from full dataset to save computation resource
}
