% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_clustering.R
\name{cytosee_clustering}
\alias{cytosee_clustering}
\alias{pro_center_calc}
\title{Opens \code{cytosee} results in an interactive session in a web browser.}
\usage{
pro_center_calc(data, label)
}
\arguments{
\item{object}{an object of \code{cytosee} class}
}
\value{
Opens a browser window with an interactive \code{shiny} app and visualize
all precomputed clusterings.
}
\description{
Opens \code{cytosee} results in an interactive session in a web browser.
}
