% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_consboost.R
\name{cytosee_consboost}
\alias{cytosee_consboost}
\title{Consboost function which using consensus clustering method}
\usage{
cytosee_consboost(object, pannal.use = NULL, cell.use = NULL, maxK = 10,
  K = 5, K_cell = 2000, reptime = 5)
}
\arguments{
\item{object}{the FCS exprs data set after transformed}

\item{K}{Number of clusters}

\item{K_cell}{The number of sampling data}

\item{reptime}{loop times for getting the best sample set}
}
\description{
Consboost function which using consensus clustering method
}
