% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_clustering.R
\name{cytosee_DensityCut}
\alias{cytosee_DensityCut}
\title{The densityCut algorithm}
\usage{
cytosee_DensityCut(object, K, knn.index, knn.dist, V, D, G, threshold,
  alpha = 0.9, adjust = TRUE, maxit = 50, eps = 1e-05, col = NULL,
  debug = FALSE, xlab = TRUE, text = NULL, ...)
}
\arguments{
\item{object}{an object of \code{cytosee} class.}

\item{K}{A integer to specify the number of neighbours in building the Knn graph.
Default to \eqn{K=\log_2(N)}, where N is the number of data points}

\item{knn.index}{An N*K data matrix for the nearest neighbour indices}

\item{knn.dist}{An N*K data matrix for the nearest neighbour distances}

\item{V}{The initial density vector of length N}

\item{D}{The dimensionality of data}

\item{G}{A sparse Knn graph, reseaved for extension}

\item{threshold}{A number between 0 and 1 specifying the saliency index to cut the tree.
If not specified, it is selecting by stability analysis of the clustering tree}

\item{alpha}{The damping factor between 0 and 1, default to 0.90}

\item{adjust}{Lotical, whether to ajdust valley height or not}

\item{maxit}{The maximum number of iteration allowed in density refinement, default to 50}

\item{eps}{The threshold in density refinement, default to 1e-5}

\item{col}{A vector of clours}

\item{debug}{Logical, whether to print debug information}

\item{xlab}{Logical, whether to show the xlab}

\item{text}{subplot label}

\item{...}{Reserved for extension}

\item{X}{A data matrix (columns are features and rows are data points)}

\item{The}{damping factor between 0 and 1, default to 0.90}

\item{nu}{The saliency index in merging trees, default to \eqn{seq(0.0, 1.0, by=0.05)}}

\item{show.plot}{Logical, whether to draw clustering results}

\item{show.tip.label}{Logical, whether to draw the tip labels of trees}
}
\description{
densityCut first roughly estimates the densities
of data points from a K-nearest neighbour graph and then refines the densities via a random
walk. A cluster consists of points falling into the basin of attraction of an estimated mode of the
underlining density function. A post-processing step merges clusters and generates a hierarchical
cluster tree. The number of clusters is selected from the most stable clustering in the hierarchical
cluster tree. Experimental results on ten synthetic benchmark datasets and two microarray gene
expression datasets demonstrate that densityCut performs better than state-of-the-art algorithms
for clustering biological datasets.
}
