% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_PhenoCL.R
\name{cytosee_DMT}
\alias{cytosee_DMT}
\title{build Divisive marker tree(DMT)

description
This is a function for building divisive marker tree. It can be used to divide the the cell populations into their own clusters by surface markers.}
\usage{
cytosee_DMT(data)
}
\arguments{
\item{data}{this data should be a dataframe which contain the all elements from the expression data of .fcs file and it's first column should be the cluster num of this event.}
}
\value{
DMT TREE list which is a tree list contain elements like "label,size,id,previd,clus"
}
