% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_DR.R
\name{cytosee_DR}
\alias{cytosee_DR}
\title{Reduce dimension of high dimension FCS data}
\usage{
cytosee_DR(data, n_core = NULL, sgd_batches = 0.5, tsne_pca = TRUE,
  method = "PCA")
}
\arguments{
\item{n_core}{how many cores you want to use}

\item{sgd_batches}{input parameter for largeVis (refer largeVis for details)}

\item{tsne_pca}{whether use pca in tsne}

\item{object}{cytosee object}
}
\description{
We apply dimension reduction to those matrix which has more than paramenters
\code{t-SNE},\code{FIt-SNE} and \code{SOM} was used
}
