% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_clustering.R
\name{cytosee_flowMeans}
\alias{cytosee_flowMeans}
\title{The flowMeans algorithm}
\usage{
cytosee_flowMeans(object, varNames = NULL, MaxN = NA, NumC = NA,
  iter.max = 50, nstart = 10, Mahalanobis = TRUE, Standardize = TRUE,
  Update = "Mahalanobis", OrthagonalResiduals = TRUE, MaxCovN = NA,
  MaxKernN = NA, addNoise = TRUE)
}
\arguments{
\item{object}{An object of \code{cytosee} class.}

\item{varNames}{A character vector specifying the variables (columns) to be included in clustering.
When it is left unspecified, all the variables will be used.}

\item{MaxN}{Maximum number of clusters. If set to NA (default) the value will be estimated automatically.}

\item{NumC}{Number of clusters. If set to NA (default) the value will be estimated automatically.}

\item{iter.max}{The maximum number of iterations allowed.}

\item{nstart}{The number of random sets used for initialization.}

\item{Mahalanobis}{Boolean value. If TRUE (default) mahalanobis distance will be used. Otherwised, euclidean distance will be used.}

\item{Standardize}{Boolean value. If TRUE (default) the data will be transformed to the [0,1] interval.}

\item{Update}{String value. If set to "Mahalanobis" the distance function will be updated at each merging iteration with recalculating mahalanobis distances.
If set to "Mean" the distance matrix will be updated after each merging step with averaging. If set to "None" the distance matrix will not be updated.}

\item{OrthagonalResiduals}{Boolean value, indicates if the residuals must be transformed to orthagonal distance or not.}

\item{MaxCovN}{Maximum number of points, used for calculating the covariance. If set to NA (default), all the points will be used.}

\item{MaxKernN}{Maximum number of points, used for counting the modes using kernel density estimation. If set to NA (default), all the points will be used.}

\item{addNoise}{Boolean value. Determines if uniform noise must be added to the data to prevent singularity issues or not.}
}
\description{
Finds a good fit to the data using k-means clustering algorithm.
Then merges the adjacent dense spherical clusters to find non-spherical clusters.
}
