% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_clustering.R
\name{cytosee_flowSOM}
\alias{cytosee_flowSOM}
\title{Run the flowSOM algorithm}
\usage{
cytosee_flowSOM(object, K, nClus = NULL,
  method = "metaClustering_consensus")
}
\arguments{
\item{object}{an object of \code{cytosee} class.}

\item{K}{Maximum number of clusters to try out.}

\item{nClus}{Exact number of clusters to use. If not NULL, max will be ignored.}

\item{method}{Clustering method to use, given as a string. Options are metaClustering_consensus,metaClustering_hclust, metaClustering_kmeans,metaClustering_som}
}
\description{
Method to run general FlowSOM workflow. Will scale the data and uses consensus meta-clustering.
}
