% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_shiny.R
\name{cytosee_go}
\alias{cytosee_go}
\title{Opens \code{SC3} results in an interactive session in a web browser.}
\usage{
cytosee_go(object)
}
\arguments{
\item{object}{an object of \code{SCESet} class}
}
\value{
Opens a browser window with an interactive \code{shiny} app and visualize
all precomputed clusterings.
}
\description{
Runs interactive \code{shiny} session of \code{SC3} based on precomputed clusterings.
}
