% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_shiny.R
\name{cytosee_gui}
\alias{cytosee_gui}
\title{Opens \code{CytoSEE} workflow in an interactive session in a web browser.}
\usage{
cytosee_gui(object)
}
\arguments{
\item{object}{an object of \code{cytosee} class}
}
\value{
Opens a browser window with an interactive \code{shiny} app and visualize
all precomputed clusterings.
}
\description{
Runs interactive \code{shiny} session of \code{CytoSEE}
}
