% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_clustering.R
\name{cytosee_SamSPECTRAL}
\alias{cytosee_SamSPECTRAL}
\title{SamSPECTRAL image based funciton for fcs data clustering.}
\usage{
cytosee_SamSPECTRAL(object, sigma, separation.factor,
  number.of.clusters = "NA", precision = 6, stabilizer = 100,
  k.for_kmeans = NA)
}
\arguments{
\item{object}{an object of \code{cytosee} class.}

\item{separation.factor}{This threshold controls to what extend clusters should be combined or kept separate.Normally,
an appropriate value will fall in range 0.3-2.}

\item{number.of.clusters}{The default value is "NA" which leads to computing the number of spectral clusters automatically,
otherwise it can be a vector of integers each of which determines the number of spectral clusters.
The output will contain a clustering resulting from each value.}

\item{sigmal}{A scaling parameter that determines the "resolution" in the spectral clustering stage.
By increasing it, more spectral clusters are identified. This can be useful when "small" population are aimed.
See the user manual for a suggestion on how to set this parameter using the eigenvalue curve.}
}
\value{
A object of cytosee with clustering result.
}
\description{
SamSPECTRAL first builds the communities to sample the data points. Then, it builds a graph and after weighting
the edges of the graph by conductance computation, it is passed to a classic spectral clustering algorithm to find
the spectral clusters. The last stage of SamSPECTRAL is to combine the spectral clusters.
The resulting "connected components" estimate biological cell populations in the data sample.
}
