% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoCL.R
\name{DMT}
\alias{DMT}
\title{build Divisive marker tree}
\usage{
DMT(data)
}
\arguments{
\item{data}{this data should be a dataframe which contain the all elements from the expression data of .fcs file and it's first column should be the cluster num of this event.}
}
\value{
DMT TREE list which is a tree list contain elements like "label,size,id,previd,clus"
}
\description{
description
This is a function for building divisive marker tree. It can be used to divide the the cell populations into their own clusters by surface markers.
}
