% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_FIt-sne.R
\name{fftRtsne}
\alias{fftRtsne}
\title{Reduce dimension of high dimension FCS data}
\usage{
fftRtsne(X, dims = 2, perplexity = 30, theta = 0.5,
  check_duplicates = TRUE, max_iter = 1000, fft_not_bh = TRUE,
  ann_not_vptree = TRUE, stop_early_exag_iter = 250,
  exaggeration_factor = 12, no_momentum_during_exag = FALSE,
  start_late_exag_iter = -1, late_exag_coeff = 1, mom_switch_iter = 250,
  momentum = 0.5, final_momentum = 0.8, learning_rate = 200,
  n_trees = 50, search_k = -1, rand_seed = -1, nterms = 3,
  intervals_per_integer = 1, min_num_intervals = 50, K = -1,
  sigma = -30, initialization = NULL, data_path = NULL,
  result_path = NULL, load_affinities = NULL, fast_tsne_path = NULL,
  nthreads = 0, perplexity_list = NULL, get_costs = FALSE, df = 1, ...)
}
\arguments{
\item{X}{data matrix}
}
\description{
We apply dimension reduction to those matrix which has more than paramenters
}
