% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_preprocess.R
\name{initiflow}
\alias{initiflow}
\title{Initialize and setup the iflow obj}
\usage{
initiflow(fcs.data, projectname = "cytoseeProject")
}
\arguments{
\item{fcs.data}{Raw input data}

\item{project}{Project name (string)}

\item{transform_method}{Method for cell normalization. Default is no normalization.}
}
\value{
Returns a cytosee object with the raw data stored in object@fcs.data.
}
\description{
Initializes the cytosee object and some optional filtering
}
\examples{
iflowobj <- initiflow(fcs.data = fcsname)
iflowobj


}
