% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoCL.R
\name{LocCL}
\alias{LocCL}
\title{It will query the possible result from local ontology datatable
description}
\usage{
LocCL(MarkerList, MaxHitsPht = 10)
}
\arguments{
\item{MarkerList}{a string word such as "CD19-CD3+CD49b++Sca1-".}

\item{MaxHitsPht}{a numerical value, the number of max possible value.}
}
\value{
if plot is TRUE ,The plot list will be returned, else a list of top 5 results and their score will be returned
}
\description{
It will query the possible result from local ontology datatable
description
}
