% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitycut.R
\name{NeatPlot}
\alias{NeatPlot}
\title{Reset the default parameters for the plot function}
\usage{
NeatPlot(..., xlab, ylab, xaxt = "s", yaxt = "s", xtck.label = TRUE,
  ytck.label = TRUE, cex.axis = 0.8)
}
\arguments{
\item{...}{go to plot}

\item{xlab}{A title for the x axis}

\item{ylab}{A title for the y axis}

\item{xaxt}{see par}

\item{yaxt}{see par}

\item{xtck.label}{Logical, wheter to draw the x axis tick labels}

\item{ytck.label}{Logical, wheter to draw the y axis tick labels}

\item{cex.axis}{cex for axis}
}
\description{
Reset the default parameters for the plot function
}
