% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoCL.R
\name{PhenoCL}
\alias{PhenoCL}
\title{Automate Query phenotypes from Cell ontology}
\usage{
PhenoCL(data, ClusterID, MarkerName = c(), MaxLabelnum = 10)
}
\arguments{
\item{data}{A matrix contain fcs data .}

\item{ClusterID}{A data.frame contain the Cluster ID of each events in data.}

\item{MarkerName}{character vectors for revising all the name of markers. Default is a empty matrix.}

\item{MaxLabelnum}{A numbmic vector determine how many top labels will be display. Default is 10.}
}
\value{
A list of all most possible results of clusters.
}
\description{
This function will automatically produce Phenotypes for each clusters and query the most possible result for them.
}
\examples{
# load the library
library(cytosee)
data(GvHD001)
data=GvHD001$data[,3,6]
ClusterID=GvHD001$label
MarkerName=c("CD4","CD8b","CD8","CD3")
result=PhenoCL(data,ClusterID,MarkerName)

}
