% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_core.R
\name{pro_reduce_dim}
\alias{pro_reduce_dim}
\title{Reduce dimension of high dimension FCS data}
\usage{
pro_reduce_dim(data, n_core = NULL, sgd_batches = 0.5, tsne_pca = TRUE,
  method = "ALL")
}
\arguments{
\item{n_core}{how many cores you want to use}

\item{sgd_batches}{input parameter for largeVis (refer largeVis for details)}

\item{tsne_pca}{whether use pca in tsne}

\item{object}{cytosee object}
}
\description{
We apply dimension reduction to those matrix which has more than paramenters
\code{t-SNE},\code{largeVis} and \code{SOM} was used
}
