% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytosee_consboost.R
\name{cytosee_consboost}
\alias{cytosee_consboost}
\title{Consboost function which using consensus clustering method}
\usage{
cytosee_consboost(object, K, sample_n = 3000, n_core = 3)
}
\arguments{
\item{object}{the FCS exprs data set after transformed}

\item{K}{Number of final clusters}

\item{sample_n}{The number of sampling data}

\item{n_core}{CPU thread use}
}
\description{
Consboost function which using consensus clustering method
}
